Restructuring of the models and the training process is done
