+++ .github/workflows/build.yml
@@ -3,6 +3,7 @@ on:
   push:
     branches:
       - main
+      - arch-changes
   pull_request:
     types: [opened, synchronize, reopened]
 jobs:
