+++ src/models/Embedding.py
@@ -1,16 +1,10 @@
 from typing import List
-
-import numpy as np
-import pandas as pd
-
-from tensorflow import keras
-from models.Triplet import Triplet
+from src.models.base.Model import Model
+from src.models.data_processing.TokenFeatures import TokenFeatures
 from tensorflow.keras import layers, regularizers, models
-from sklearn.model_selection import train_test_split
-from src.data_processing.commons import std_initial_preprocess
+from tensorflow import keras
 
-
-class Embedding(Triplet):
+class Embedding(TokenFeatures, Model):
 
     def __init__(self,
                  input_size: int = 100,
@@ -19,24 +13,20 @@ class Embedding(Triplet):
                  max_val: int = 99756 + 1):
 
         self.max_val = max_val
-        super().__init__(name="embedding",
-                         input_size=input_size, output_size=output_size,
-                         make_initial_preprocess=make_initial_preprocess)
+        self.output_size = output_size
 
-    def create_model(self,
-                     activation: str = "linear",
-                     L2_lambda: float = 0.02,
-                     conv_sizes: List[int] = [2, 4, 16],
-                     emb_height: int = 100):
+        Model.__init__(self)
+        TokenFeatures.__init__(self, name="embedding",
+                               input_size=input_size,
+                               make_initial_preprocess=make_initial_preprocess)
+        self.model = self.create_model()
 
-        conv_channels = 1
-        input_layer = layers.Input(shape=(self.input_size, 1))
-
-        embeddings = layers.Embedding(self.max_val, emb_height,
-                                      mask_zero=True, input_length=self.input_size)(input_layer)
-
-        reshape1 = layers.Reshape((self.input_size, emb_height, 1))(embeddings)
-
+    def create_after_emb(self, reshape1,
+                         conv_channels=1,
+                         emb_height=100,
+                         activation="linear",
+                         L2_lambda=0.02,
+                         conv_sizes=[2, 4, 16]):
         # parallel piece
         convolutions = [layers.Conv2D(conv_channels, (conv_size, emb_height),
                                       padding="same", activation=activation,
@@ -64,41 +54,25 @@ class Embedding(Triplet):
         norm1 = layers.LayerNormalization(axis=-1)(flatten)
         drop2 = layers.Dropout(0.5)(norm1)
         dense = layers.Dense(self.output_size)(drop2)
+        return dense
+
+    def create_model(self,
+                     activation: str = "linear",
+                     L2_lambda: float = 0.02,
+                     conv_sizes: List[int] = [2, 4, 16],
+                     emb_height: int = 100):
+
+        conv_channels = 1
+        input_layer = layers.Input(shape=(self.input_size, 1))
+
+        embeddings = layers.Embedding(self.max_val, emb_height,
+                                      mask_zero=True, input_length=self.input_size)(input_layer)
+
+        reshape1 = layers.Reshape((self.input_size, emb_height, 1))(embeddings)
+
+        dense = self.create_after_emb(reshape1, conv_channels, emb_height, activation, L2_lambda, conv_sizes)
         result =  models.Model(input_layer, dense)
 
         print(result.summary())
         keras.utils.plot_model(result, "{}.png".format(self.name), show_shapes=True)
-
         return result
-
-    @staticmethod
-    def crop_to(X: np.ndarray,
-                y: np.ndarray,
-                crop: int = 100,
-                threshold: int = 80):
-
-        new_X = []
-        new_y = []
-        for old_x, old_y in zip(X, y):
-            for el in old_x.reshape(-1, crop):
-                if np.count_nonzero(el) > threshold * crop // 100:
-                    new_X.append(list(el))
-                    new_y.append(old_y)
-
-        new_X = np.array(new_X).reshape((-1, crop, 1))
-        new_y = np.array(new_y)
-        return new_X, new_y
-
-    def initial_preprocess(self, df_path: str, tmp_dataset_filename: str):
-        std_initial_preprocess(self.input_size, df_path, tmp_dataset_filename)
-
-    def secondary_preprocess(self, tmp_dataset_filename: str):
-        dataset = pd.read_json(tmp_dataset_filename)
-
-        X = dataset.tokens.values
-        X = np.array(list(X)).reshape((-1, self.input_size))
-
-        y = np.array(dataset.username)
-        X, y = self.crop_to(X, y, 100)
-        X_train, X_test, y_train, y_test = train_test_split(X, y)
-        return X_train, X_test, y_train, y_test
