+++ src/visualization/VisualizerCharFeatures.py
@@ -1,0 +1,8 @@
+import unittest
+
+class MyTestCase(unittest.TestCase):
+    def test_something(self):
+        self.assertEqual(True, False)
+
+if __name__ == '__main__':
+    unittest.main()
